app/requirements.txt
djangorestframework==3.14.0
drf-yasg
coreapi
django-cors-headers
psycopg2-binary
beautifulsoup4
requests